requirements.txt
PyMuPDF
langchain
openai
faiss-cpu
python-dotenv
streamlit
langchain-openai
tiktoken
langchain-community


